"Update package-lock.json, package.json and index.js to add new dependencies and routes; remove serjer.js"
This commit updates the package-lock.json and package.json files to add new dependencies such as cookie-parser, dotenv, envalid, jsonwebtoken, mongodb, morgan, and swagger-ui-express. It also removes the serjer.js file and updates the index.js file to import and use the new dependencies and routes. The dev script in package.json is also updated to run nodemon on the server/server.js file instead of the previous index.js file.